feat: KafkaInstances components to render the list of Kafka instances
This is a full view component, it's meant to be used as the direct child of a `Page` component as it provides all the expected visuals for this page.

Some components have been moved and renamed as they will be used by this new page component and are not really meant to be used by other modules.